+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/supplies/DeleteDto.kt
@@ -1,0 +1,8 @@
+package com.app.arcabyolimpo.data.remote.dto.supplies
+
+import com.google.gson.annotations.SerializedName
+
+// You only define the data parameters
+data class DeleteDto (
+    @SerializedName("id") val id: String
+)